requirements.txt
fastapi
tensorflow
numpy
Pillow
python-multipart
uvicorn
starlette
authlib
httpx
itsdangerous
SQLAlchemy
psycopg